add program_invocation_name.c
This one is used by kmod as a technique to dynamically call
its operational functions at runtime depends on which
symlinked file user executed.

In which all of symlinked files link back to the same
executable binary called kmod.

Note also for the #define of _GNU_SOURCE needs to be there,
as well as inclusion of errno.h if necessary before other
inclusion of main header files.